requirements.txt
asgiref==3.7.2
dj-static==0.0.6
Django==5.0
django-stdimage==6.0.2
gunicorn==21.2.0
packaging==23.2
Pillow==10.1.0
psycopg2-binary==2.9.9
sqlparse==0.4.4
static3==0.7.0
